Move extensions to separate package
